Complete Home api

+++ database/migrations/2024_01_23_102325_create_product_ratings_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreign('user_id')->references('id')->on('users');
             $table->unsignedBigInteger('product_id');
             $table->foreign('product_id')->references('id')->on('products');
-            $table->unsignedBigInteger('reviews');
+            $table->decimal('reviews',10,2);
             $table->longText('add_info');
             $table->timestamps();
         });
